% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{prop_01}
\alias{prop_01}
\title{p_01 estimator}
\usage{
prop_01(x, y)
}
\arguments{
\item{x, y}{vectors to be correlated. Must be numeric.}
}
\value{
p_01 estimator
}
\description{
computes estimate of parameter p_01 based on sample proportions.
}
\keyword{internal}
