% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{prop_10}
\alias{prop_10}
\title{p_10 estimator}
\usage{
prop_10(x, y)
}
\arguments{
\item{x, y}{vectors to be correlated. Must be numeric.}
}
\value{
p_10 estimator
}
\description{
computes estimate of parameter p_01 based on sample proportions.
}
\keyword{internal}
