% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{prop_11}
\alias{prop_11}
\title{p_11 estimator}
\usage{
prop_11(x, y)
}
\arguments{
\item{x, y}{vectors to be correlated. Must be numeric.}
}
\value{
p_11 estimator
}
\description{
computes estimate of parameter p_11 based on sample proportions.
}
\keyword{internal}
